Error fixup and password checking

+++ frontend/src/pages/Register.tsx
@@ -1,6 +1,6 @@
 import { useNavigate } from "react-router-dom";
 import { Box, TextField, Button, Typography } from "@mui/material";
-import { useState } from "react";
+import { useContext, useState, type FormEvent } from "react";
 import { apiRegister } from "../api";
 import ErrorContext from "../context/ErrorContext";
 
@@ -14,6 +14,11 @@ const Register = () => {
   const navigate = useNavigate();
 
   const handleRegister = async () => {
+    if (password !== confirmPassword) {
+      showError("Passwords do not match");
+      return;
+    }
+
     try {
       const token = await apiRegister(name, email, password);
       localStorage.setItem("token", token);
@@ -23,13 +28,15 @@ const Register = () => {
     }
   };
 
+  const handleSubmit = (e: FormEvent) => {
+    e.preventDefault();
+    handleRegister();
+  };
+
   return (
     <Box
       component="form"
-      onSubmit={(e) => {
-        e.preventDefault();
-        handleRegister();
-      }}
+      onSubmit={handleSubmit}
       className="flex flex-col items-center justify-center h-screen gap-4"
     >
       <Typography variant="h5" fontWeight="bold">
@@ -63,7 +70,7 @@ const Register = () => {
         type="password"
         onChange={(e) => setConfirmPassword(e.target.value)}
       />
-      <Button variant="contained" onClick={handleRegister}>
+      <Button type="submit" variant="contained">
         Register
       </Button>
     </Box>
